refactor: make main function more readable
Extracted functions to install poetry and dependencies
